feat: v2 management api group add action resource instance filter
