+++ saas/backend/apps/role/serializers.py
@@ -444,9 +444,3 @@ class SubsetMangerDetailSLZ(GradeMangerDetailSLZ):
             "subject_scopes",
             "sync_perm",
         )
-
-
-class ThinGradeManagerSLZ(serializers.ModelSerializer):
-    class Meta:
-        model = Role
-        fields = ("id", "name", "description")
